PYIC-1718: Rename the cri start lambda
