+++ lambdas/build-cri-oauth-request/src/main/java/uk/gov/di/ipv/core/buildcrioauthrequest/domain/CriDetails.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.core.credentialissuer.domain;
+package uk.gov.di.ipv.core.buildcrioauthrequest.domain;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
