+++ lambdas/build-cri-oauth-request/src/main/resources/IpvLambdaJsonLayout.json
