+++ lambdas/build-cri-oauth-request/src/main/resources/log4j2.yaml
