+++ settings.gradle
@@ -10,7 +10,7 @@ include "lib",
 		"lambdas:sessionstart",
 		"lambdas:useridentity",
 		"lambdas:process-journey-step",
-		"lambdas:credentialissuerstart",
+		"lambdas:build-cri-oauth-request",
 		"lambdas:credentialissuererror",
 		"lambdas:validatecricheck",
 		"integration-test"
